+++ pdf/lib/pdf.dart
@@ -37,7 +37,6 @@ part 'src/border.dart';
 part 'src/catalog.dart';
 part 'src/color.dart';
 part 'src/colors.dart';
-part 'src/compatibility.dart';
 part 'src/data_types.dart';
 part 'src/document.dart';
 part 'src/encryption.dart';